ContosoGraphSummarization_git/Contoso_Summarization_Tool/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiolimiter==1.1.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
anytree==2.12.1
asgiref==3.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.3
attrs==24.2.0
autograd==1.7.0
azure-ai-documentintelligence==1.0.0b4
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-search-documents==11.5.1
azure-storage-blob==12.22.0
beartype==0.18.5
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
cloudpickle==3.0.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
configobj==5.0.8
configparser==7.1.0
contourpy==1.3.0
cramjam==2.8.3
cryptography==43.0.1
cycler==0.12.1
dask==2024.8.2
dask-expr==1.1.13
dataclasses-json==0.6.7
datashaper==0.0.49
debugpy @ file:///C:/b/abs_c0y1fjipt2/croot/debugpy_1690906864587/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
deprecation==2.1.0
devtools==0.12.2
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docx2txt==0.8
environs==11.0.0
etelemetry==0.3.1
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
fastparquet==2024.5.0
filelock==3.16.0
fitz==0.0.1.dev2
FLAML==2.3.1
Flask==3.0.3
Flask-Markdown==0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
gensim==4.3.3
graphrag==0.3.6
graspologic==3.4.1
graspologic-native==1.2.1
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
hyppo==0.4.0
idna==3.8
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/ipython_1725050320818/work
isodate==0.6.1
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json_repair==0.28.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1716472197302/work
jupyter_core @ file:///D:/bld/jupyter_core_1710257272359/work
kiwisolver==1.4.7
lancedb==0.13.0
langchain==0.3.1
langchain-community==0.3.0
langchain-core==0.3.6
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langgraph==0.2.28
langgraph-checkpoint==1.0.9
langsmith==0.1.128
lightgbm==4.5.0
linkify-it-py==2.0.3
llvmlite==0.43.0
locket==1.0.0
looseversion==1.3.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdit-py-plugins==0.4.2
mdurl==0.1.2
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
mypy-extensions==1.0.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.48.0
orjson==3.10.7
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.2
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
partd==1.4.2
pathlib==1.0.1
patsy==0.5.6
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==10.4.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1725821724470/work
portalocker==2.10.1
POT==0.9.4
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1718047967974/work
prov==2.0.1
psutil @ file:///C:/Windows/Temp/abs_b2c2fd7f-9fd5-4756-95ea-8aed74d0039flsd9qufz/croots/recipe/psutil_1656431277748/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
py==1.11.0
pyaml-env==1.2.1
pyarrow==15.0.0
pyautogen==0.3.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydot==3.0.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
PyJWT==2.9.0
pylance==0.17.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pynndescent==0.5.13
pypandoc==1.13
pyparsing==3.1.4
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pywin32==305.1
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1666828590571/work
ratelimiter==1.2.0.post0
rdflib==6.3.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.8.1
rpds-py==0.20.0
scikit-learn==1.5.2
scipy==1.12.0
seaborn==0.13.2
simplejson==3.19.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
statsmodels==0.14.2
swifter==1.4.0
tenacity==8.5.0
termcolor==2.4.0
textual==0.70.0
threadpoolctl==3.5.0
tiktoken==0.7.0
toolz==0.12.1
tornado @ file:///D:/bld/tornado_1666788744359/work
tqdm==4.66.5
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
traits==6.3.2
typing-inspect==0.9.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.1
uc-micro-py==1.0.3
umap-learn==0.5.6
urllib3==2.2.3
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==3.0.4
wrapt==1.16.0
xgboost==2.1.1
yarl==1.11.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1724730934107/work
